requirements.txt
flask==3.1.1
flask-cors==6.0.0
gunicorn==23.0.0
aiohttp==3.9.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.24.3
pickle-mixin==1.0.2
requests==2.32.3